+++ pyffi/formats/nif/__init__.py
@@ -3209,6 +3209,18 @@ class NifFormat(FileFormat):
             self.dimensions.y *= scale
             self.dimensions.z *= scale
 
+    class BSDismemberSkinInstance:
+        def get_dismember_partitions(self):
+            """Return triangles and body part indices."""
+            triangles = []
+            trianglepartmap = []
+            for bodypart, skinpartblock in zip(
+                self.partitions, self.skin_partition.skin_partition_blocks):
+                part_triangles = list(skinpartblock.get_mapped_triangles())
+                triangles += part_triangles
+                trianglepartmap += [bodypart.body_part] * len(part_triangles)
+            return triangles, trianglepartmap
+
     class ControllerLink:
         """
         >>> from pyffi.formats.nif import NifFormat