+++ makezip.sh
@@ -22,7 +22,7 @@ fi
 rm MANIFEST
 python setup.py -q sdist --format=zip
 python setup.py -q sdist --format=bztar
-python setup.py --command-packages bdist_nsi bdist_nsi --bitmap=win-install/pyffi_install_164x314.bmp --headerbitmap=win-install/pyffi_install_150x57.bmp --run2to3 --msvc2008sp1 --nshextra=win-install/pyffi.nsh --target-versions=2.5,2.6,2.7,3.0,3.1,3.2 --maya --blender
+python setup.py --command-packages bdist_nsi bdist_nsi --bitmap=win-install/pyffi_install_164x314.bmp --headerbitmap=win-install/pyffi_install_150x57.bmp --msvc2008sp1 --nshextra=win-install/pyffi.nsh --target-versions=2.5,2.6,2.7 --maya --blender
 
 version=`python setup.py -V`
 wcrev=`git log -1 --pretty=format:%h`