+++ CHANGELOG.rst
@@ -26,6 +26,9 @@ Release 2.1.7 (in development)
 * Fix optimize spell when no vertices are left after removing duplicate
   vertices and degenerate triangles (reported by Gratis_monsta).
 
+* Fixed tangent space issue along uv seams (reported by Gratis_monsta
+  and Tommy_H, see issue #3120585).
+
 * Log an error instead of raising an exception on invalid enum values
   (fixes issue #3127161, reported by rlibiez).